\name{NO.PING.PONG-package}

\alias{NO.PING.PONG-package}

\docType{package}

\title{NO.PING.PONG}

\description{This package contains two functions, NO.PING.PONG and PLOT_NO.PING.PONG, 
	for revealing what happens when effect size estimates from previous studies 
	are taken into account when evaluating each new dataset in a study sequence. 
	The analyses can be conducted for cumulative meta-analyses and for 
	Bayesian data analyses. The findings from these methods can be contrasted with
	those from null hypothesis statistical sigificance testing, which does
	not take previous findings into account.
	
	The package contains sample data for a wide selection 
	of research topics that can be used to illustrate why previous findings 
	should be taken into account when attempting to reach a conclusion about an 
	effect when evaluating a new dataset. Jointly considering previous findings 
	along with new data is more likely to result in correct conclusions than does 
	the traditional practice of not incorporating previous findings and which often
	results in a back and forth ping-pong of conclusions when evaluating
	a sequence of studies.
}
\references{
       {O'Connor, B. P., & Ermacora, D. (2021). Unnecessary ping-pong: Illustrations of 
       	why previous findings should be taken into account when evaluating new datasets. 
       	\emph{Canadian Journal of Behavioural Science, 53(3),} 328-341. https://doi.org/10.1037/cbs0000259}       	
\cr\cr {O'Connor, B. P., & Khattar, N. (2022). Controversies regarding 
	    null hypothesis testing. In W. O'Donohue, A. Masuda, & S. O. Lilienfeld (Eds.). 
	    \emph{Avoiding Questionable Research Practices in Applied Psychology} (pp. 147-174). 
	    Cham, Switzerland: Springer Nature Switzerland.}
}    
	    
	    
